Signout API and product and order layout created

+++ admin-app/src/actions/auth.actions.js
@@ -7,7 +7,9 @@ export const login = (user) => {
 
     return async ( dispatch ) => {
 
-        dispatch({ type: authConstants.LOGIN_REQUEST });
+        dispatch({ 
+            type: authConstants.LOGIN_REQUEST 
+        });
         const res = await axios.post(`/admin/signin`, {
             ...user
         });
@@ -17,7 +19,7 @@ export const login = (user) => {
             localStorage.setItem('token', token);
             localStorage.setItem('user', JSON.stringify(user));
             dispatch({
-                type: authConstants.LOGIN_SUCCES,
+                type: authConstants.LOGIN_SUCCESS,
                 payload: {
                     token, user
                 }
@@ -32,19 +34,17 @@ export const login = (user) => {
     }
 }
 
-
-
 export const isUserLoggedIn = () => {
     return async dispatch => {
         const token = localStorage.getItem('token');
         if(token){
           const user = JSON.parse(localStorage.getItem('user'));
           dispatch({
-            type: authConstants.LOGIN_SUCCES,
-            payload: {
-                token, user
-            }
-        }); 
+                type: authConstants.LOGIN_SUCCESS,
+                payload: {
+                    token, user
+                }
+            }); 
         }else{
             dispatch({
                 type: authConstants.LOGIN_FAILURE,
@@ -56,9 +56,23 @@ export const isUserLoggedIn = () => {
 
 export const signout = () => {
     return async dispatch => {
-        localStorage.clear();
-        dispatch({
-            type: authConstants.LOGOUT_REQUEST
-        })
+
+        dispatch({ type: authConstants.LOGOUT_REQUEST });
+
+        const res = await axios.post('/admin/signout')
+
+        if(res.status === 200){
+            localStorage.clear();
+            dispatch({
+                type: authConstants.LOGOUT_SUCCESS
+            })
+        }else{
+            dispatch({
+                type: authConstants.LOGOUT_FAILURE,
+                payload: { error: res.data.error }
+            })
+        }
+        
+        
     }
 }
